requirements.txt
numpy==1.23.5
Pillow==9.5.0
requests==2.25.1
pymongo==3.11.3
Flask==2.3.0
pytest==7.4.2
psycopg2==2.9.7
python-dotenv==1.0.0
coverage==7.3.1
pytest-pycodestyle==2.3.1
pytest-flakes==4.0.5
pytest-cov==4.1.0
